requirements.txt
flask
tensorflow
numpy
pillow
opencv-python
matplotlib
scikit-learn
werkzeug
